pyproject.toml
[tool.poetry]
name = "careespace"
version = "0.1.0"
description = ""
authors = ["arybjan"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
